test: Improve snapshot infrastructure and organization (#687)
* test: Add helper to enable chromatic snapshots

* test(button): Refactor visual testing stories to use CSF

* test(form-field): Refactor visual testing stories to use CSF

* test(pagination): Refactor visual testing stories to use CSF

* test(select): Refactor visual testing stories to use CSF

* test(popup): Refactor visual testing stories to use CSF

* test(toast): Refactor visual testing stories to use CSF

* test(form-field): Rename testing stories into their own section

* test(button): Rename testing stories into their own section

* test: Rename remainder of testing stories into their own section

* fix(form-field): Fix import

* refactor: Add category to create-component script and use CSF

* feat: Add CSF visual testing story to create-component script

* fix(form-field): Remove unused variable

* fix(form-field): Properly export theme stories

* refactor: Use a decorator for enabling snapshots

* feat: Add override to withSnapshotsEnabled so it can be used per story
