+++ modules/_labs/pagination/react/stories/stories.tsx
@@ -4,7 +4,6 @@ import {boolean, number, text} from '@storybook/addon-knobs';
 import {storiesOf} from '@storybook/react';
 import React, {useEffect, useState} from 'react';
 import withReadme from 'storybook-readme/with-readme';
-import {ComponentStatesTable} from '../../../../../utils/storybook';
 
 import {Pagination} from '@workday/canvas-kit-labs-react-pagination';
 import README from '../README.md';
@@ -108,47 +107,5 @@ storiesOf('Labs|Pagination/React', module)
           {...getAriaLabels()}
         />
       </Wrapper>
-    );
-  });
-
-storiesOf('Labs|Pagination/React/Visual Testing', module)
-  .addParameters({
-    chromatic: {
-      disable: false,
-    },
-  })
-  .add('States', () => {
-    const defaults = {
-      total: 1000,
-      pageSize: 10,
-      showLabel: true,
-      showGoTo: true,
-      currentPage: 1,
-      onPageChange: (_: any) => {
-        /* don't do anything */
-      },
-    };
-    return (
-      <ComponentStatesTable
-        rowProps={[
-          {label: 'Basic 3 Pages', props: {total: 30, showGoTo: false}},
-          {label: 'Basic 7 Pages', props: {total: 70, showGoTo: false}},
-          {label: 'First Page', props: {currentPage: 1}},
-          {label: 'Middle Page', props: {currentPage: 50}},
-          {label: 'Last Page', props: {currentPage: 100}},
-        ]}
-        columnProps={[
-          {
-            label: 'Desktop',
-            props: {width: 800},
-          },
-          {
-            label: 'Mobile',
-            props: {width: 499},
-          },
-        ]}
-      >
-        {props => <Pagination {...defaults} {...props} />}
-      </ComponentStatesTable>
     );
   });
