+++ modules/form-field/react/stories/stories_Checkbox.tsx
@@ -2,13 +2,7 @@
 import * as React from 'react';
 import {storiesOf} from '@storybook/react';
 import withReadme from 'storybook-readme/with-readme';
-import {StaticStates} from '@workday/canvas-kit-labs-react-core';
-import {
-  ControlledComponentWrapper,
-  ComponentStatesTable,
-  permutateProps,
-  customColorTheme,
-} from '../../../../utils/storybook';
+import {ControlledComponentWrapper} from '../../../../utils/storybook';
 
 import {Checkbox} from '../../../checkbox/react';
 import FormField from '../index';
@@ -118,70 +112,3 @@ storiesOf('Components|Inputs/Checkbox/React/Left Label', module)
       {control(<Checkbox label="Checkbox option" />)}
     </FormField>
   ));
-
-const CheckboxStates = () => (
-  <StaticStates>
-    <ComponentStatesTable
-      rowProps={permutateProps(
-        {
-          checked: [{value: true, label: 'Checked'}, {value: false, label: 'Unchecked'}],
-          indeterminate: [{value: true, label: 'Indeterminate'}, {value: false, label: ''}],
-          error: [
-            {value: undefined, label: ''},
-            {value: Checkbox.ErrorType.Alert, label: 'Alert'},
-            {value: Checkbox.ErrorType.Error, label: 'Error'},
-          ],
-        },
-        props => {
-          if (props.indeterminate && !props.checked) {
-            return false;
-          }
-          return true;
-        }
-      )}
-      columnProps={permutateProps(
-        {
-          className: [
-            {label: 'Default', value: ''},
-            {label: 'Hover', value: 'hover'},
-            {label: 'Focus', value: 'focus'},
-            {label: 'Focus Hover', value: 'focus hover'},
-            {label: 'Active', value: 'active'},
-            {label: 'Active Hover', value: 'active hover'},
-          ],
-          disabled: [{label: '', value: false}, {label: 'Disabled', value: true}],
-        },
-        props => {
-          if (props.disabled && !['', 'hover'].includes(props.className)) {
-            return false;
-          }
-          return true;
-        }
-      )}
-    >
-      {props => (
-        <Checkbox
-          {...props}
-          onChange={() => {}} // eslint-disable-line no-empty-function
-          label="Checkbox"
-        />
-      )}
-    </ComponentStatesTable>
-  </StaticStates>
-);
-
-storiesOf('Components|Inputs/Checkbox/React/Visual Testing', module)
-  .addParameters({
-    component: Checkbox,
-    chromatic: {
-      disable: false,
-    },
-  })
-  .addDecorator(withReadme(README))
-  .add('States', () => <CheckboxStates />)
-  .addParameters({
-    canvasProviderDecorator: {
-      theme: customColorTheme,
-    },
-  })
-  .add('Theming', () => <CheckboxStates />);
