+++ modules/form-field/react/stories/stories_Select.tsx
@@ -2,8 +2,7 @@
 import * as React from 'react';
 import {storiesOf} from '@storybook/react';
 import withReadme from 'storybook-readme/with-readme';
-import {StaticStates} from '@workday/canvas-kit-labs-react-core';
-import {controlComponent, ComponentStatesTable, permutateProps} from '../../../../utils/storybook';
+import {controlComponent} from '../../../../utils/storybook';
 
 import FormField from '..';
 import README from '../../../select/react/README.md';
@@ -208,54 +207,4 @@ storiesOf('Components|Inputs/Select/React/Left Label', module)
         </Select>
       )}
     </FormField>
-  ));
-
-storiesOf('Components|Inputs/Select/React/Visual Testing', module)
-  .addParameters({
-    component: Select,
-    chromatic: {
-      disable: false,
-    },
-  })
-  .addDecorator(withReadme(README))
-  .add('States', () => (
-    <StaticStates>
-      <ComponentStatesTable
-        rowProps={[
-          {label: 'Default', props: {}},
-          {label: 'Alert', props: {error: Select.ErrorType.Alert}},
-          {label: 'Error', props: {error: Select.ErrorType.Error}},
-        ]}
-        columnProps={permutateProps(
-          {
-            className: [
-              {label: 'Default', value: ''},
-              {label: 'Hover', value: 'hover'},
-              {label: 'Focus', value: 'focus'},
-              {label: 'Focus Hover', value: 'focus hover'},
-              {label: 'Active', value: 'active'},
-              {label: 'Active Hover', value: 'active hover'},
-            ],
-            disabled: [{label: '', value: false}, {label: 'Disabled', value: true}],
-          },
-          props => {
-            if (props.disabled && !['', 'hover'].includes(props.className)) {
-              return false;
-            }
-            return true;
-          }
-        )}
-      >
-        {props => (
-          <Select
-            {...props}
-            style={{minWidth: 60, width: 100}}
-            onChange={() => {}} // eslint-disable-line no-empty-function
-          >
-            <SelectOption value="email" label="E-mail" />
-            <SelectOption value="phone" label="Phone" />
-          </Select>
-        )}
-      </ComponentStatesTable>
-    </StaticStates>
   ));
