+++ modules/form-field/react/stories/stories_Switch.tsx
@@ -2,13 +2,7 @@
 import * as React from 'react';
 import {storiesOf} from '@storybook/react';
 import withReadme from 'storybook-readme/with-readme';
-import {
-  ControlledComponentWrapper,
-  ComponentStatesTable,
-  permutateProps,
-  customColorTheme,
-} from '../../../../utils/storybook';
-import {StaticStates} from '@workday/canvas-kit-labs-react-core';
+import {ControlledComponentWrapper} from '../../../../utils/storybook';
 import {Switch} from '../../../switch/react';
 import FormField from '../index';
 import README from '../../../switch/react/README.md';
@@ -95,68 +89,3 @@ storiesOf('Components|Inputs/Switch/React/Left Label', module)
       {control(<Switch />)}
     </FormField>
   ));
-
-const SwitchStates = () => (
-  <StaticStates>
-    <ComponentStatesTable
-      rowProps={permutateProps(
-        {
-          checked: [{value: true, label: 'Checked'}, {value: false, label: 'Unchecked'}],
-          error: [
-            {value: undefined, label: ''},
-            {value: Switch.ErrorType.Alert, label: 'Alert'},
-            {value: Switch.ErrorType.Error, label: 'Error'},
-          ],
-        },
-        props => {
-          if (props.indeterminate && !props.checked) {
-            return false;
-          }
-          return true;
-        }
-      )}
-      columnProps={permutateProps(
-        {
-          className: [
-            {label: 'Default', value: ''},
-            {label: 'Hover', value: 'hover'},
-            {label: 'Focus', value: 'focus'},
-            {label: 'Focus Hover', value: 'focus hover'},
-            {label: 'Active', value: 'active'},
-            {label: 'Active Hover', value: 'active hover'},
-          ],
-          disabled: [{label: '', value: false}, {label: 'Disabled', value: true}],
-        },
-        props => {
-          if (props.disabled && !['', 'hover'].includes(props.className)) {
-            return false;
-          }
-          return true;
-        }
-      )}
-    >
-      {props => (
-        <Switch
-          {...props}
-          onChange={() => {}} // eslint-disable-line no-empty-function
-        />
-      )}
-    </ComponentStatesTable>
-  </StaticStates>
-);
-
-storiesOf('Components|Inputs/Switch/React/Visual Testing', module)
-  .addParameters({
-    component: Switch,
-    chromatic: {
-      disable: false,
-    },
-  })
-  .addDecorator(withReadme(README))
-  .add('States', () => <SwitchStates />)
-  .addParameters({
-    canvasProviderDecorator: {
-      theme: customColorTheme,
-    },
-  })
-  .add('Theming', () => <SwitchStates />);
