+++ modules/form-field/react/stories/stories_TextInput.tsx
@@ -2,13 +2,7 @@
 import * as React from 'react';
 import {storiesOf} from '@storybook/react';
 import withReadme from 'storybook-readme/with-readme';
-import {StaticStates} from '@workday/canvas-kit-labs-react-core';
-import {
-  controlComponent,
-  ComponentStatesTable,
-  permutateProps,
-  customColorTheme,
-} from '../../../../utils/storybook';
+import {controlComponent} from '../../../../utils/storybook';
 
 import {TextInput} from '../../../text-input/react';
 import FormField from '../index';
@@ -158,70 +152,3 @@ storiesOf('Components|Inputs/Text Input/React/Left Label', module)
       {controlComponent(<TextInput placeholder="Placeholder" />)}
     </FormField>
   ));
-
-const TextInputStates = () => (
-  <StaticStates>
-    <ComponentStatesTable
-      rowProps={permutateProps(
-        {
-          value: [{value: 'Input value', label: 'With Value'}, {value: '', label: 'No Value'}],
-          placeholder: [{value: 'Placeholder', label: 'Placeholder'}],
-          error: [
-            {value: undefined, label: ''},
-            {value: TextInput.ErrorType.Alert, label: 'Alert'},
-            {value: TextInput.ErrorType.Error, label: 'Error'},
-          ],
-        },
-        props => {
-          if (props.value === '' && !props.placeholder) {
-            return false;
-          }
-          return true;
-        }
-      )}
-      columnProps={permutateProps(
-        {
-          className: [
-            {label: 'Default', value: ''},
-            {label: 'Hover', value: 'hover'},
-            {label: 'Focus', value: 'focus'},
-            {label: 'Focus Hover', value: 'focus hover'},
-            {label: 'Active', value: 'active'},
-            {label: 'Active Hover', value: 'active hover'},
-          ],
-          disabled: [{label: '', value: false}, {label: 'Disabled', value: true}],
-        },
-        props => {
-          if (props.disabled && !['', 'hover'].includes(props.className)) {
-            return false;
-          }
-          return true;
-        }
-      )}
-    >
-      {props => (
-        <TextInput
-          {...props}
-          style={{minWidth: 60, width: 100}}
-          onChange={() => {}} // eslint-disable-line no-empty-function
-        />
-      )}
-    </ComponentStatesTable>
-  </StaticStates>
-);
-
-storiesOf('Components|Inputs/Text Input/React/Visual Testing', module)
-  .addParameters({
-    component: TextInput,
-    chromatic: {
-      disable: false,
-    },
-  })
-  .addDecorator(withReadme(README))
-  .add('States', () => <TextInputStates />)
-  .addParameters({
-    canvasProviderDecorator: {
-      theme: customColorTheme,
-    },
-  })
-  .add('Theming', () => <TextInputStates />);
