+++ modules/form-field/react/stories/visual-testing/stories_TextInput.tsx
@@ -1,0 +1,65 @@
+/// <reference path="../../../../../typings.d.ts" />
+import * as React from 'react';
+import {StaticStates} from '@workday/canvas-kit-labs-react-core';
+import {
+  ComponentStatesTable,
+  permutateProps,
+  withSnapshotsEnabled,
+} from '../../../../../utils/storybook';
+import {TextInput} from '../../../../text-input/react';
+
+export default withSnapshotsEnabled({
+  title: 'Testing|React/Inputs/Text Input',
+  component: TextInput,
+});
+
+export const TextInputStates = () => (
+  <StaticStates>
+    <ComponentStatesTable
+      rowProps={permutateProps(
+        {
+          value: [{value: 'Input value', label: 'With Value'}, {value: '', label: 'No Value'}],
+          placeholder: [{value: 'Placeholder', label: 'Placeholder'}],
+          error: [
+            {value: undefined, label: ''},
+            {value: TextInput.ErrorType.Alert, label: 'Alert'},
+            {value: TextInput.ErrorType.Error, label: 'Error'},
+          ],
+        },
+        props => {
+          if (props.value === '' && !props.placeholder) {
+            return false;
+          }
+          return true;
+        }
+      )}
+      columnProps={permutateProps(
+        {
+          className: [
+            {label: 'Default', value: ''},
+            {label: 'Hover', value: 'hover'},
+            {label: 'Focus', value: 'focus'},
+            {label: 'Focus Hover', value: 'focus hover'},
+            {label: 'Active', value: 'active'},
+            {label: 'Active Hover', value: 'active hover'},
+          ],
+          disabled: [{label: '', value: false}, {label: 'Disabled', value: true}],
+        },
+        props => {
+          if (props.disabled && !['', 'hover'].includes(props.className)) {
+            return false;
+          }
+          return true;
+        }
+      )}
+    >
+      {props => (
+        <TextInput
+          {...props}
+          style={{minWidth: 60, width: 100}}
+          onChange={() => {}} // eslint-disable-line no-empty-function
+        />
+      )}
+    </ComponentStatesTable>
+  </StaticStates>
+);
