+++ utils/storybook/index.ts
@@ -5,6 +5,7 @@ export {
   useControlledCheck,
 } from './ControlledComponentWrapper';
 export {customColorTheme} from './customThemes';
+export {withSnapshotsEnabled} from './withSnapshotsEnabled';
 export {default as CanvasProviderDecorator} from './CanvasProviderDecorator';
 export {default as FontsDecorator} from './FontsDecorator';
 export {default as ComponentStatesTable, permutateProps} from './ComponentStatesTable';
